MSM8916: ARM: configs: Add additional config options for MSM8916 on ARM32 (v3)
The kernel configuration is largely shared with ARM64, so these
are only the additional options needed for ARM32. It needs to be
concatenated to the ARM64 msm8916_defconfig (see comment).

v2: Disable CONFIG_ARM_SMMU on ARM32 since in practice it's only useful
with 64-bit custom TZ.
v3: CONFIG_ARM_PSCI=y for DB410c