FIXME: soc: qcom: Add MSM8916/Cortex-A CPU/SMP operations
This was originally based on
https://lore.kernel.org/linux-arm-msm/1429041520-6720-3-git-send-email-galak@codeaurora.org/

But now the CPU core bring-up sequence is more taken from
https://lore.kernel.org/linux-arm-msm/20210513153442.52941-3-bartosz.dudziak@snejp.pl/
(with missing iounmap() added).

My plan is to get rid of all of this, and bring up the cores in lk2nd
instead. This can then be used with SMP spin table without any changes
in ARM64 core code.